refactor: Add `helpers` subdirectory
And remove `existsOne` from the adapter.
