+++ src/pseudo-selectors/filters.ts
@@ -1,6 +1,6 @@
 import getNCheck from "nth-check";
 import * as boolbase from "boolbase";
-import { cacheParentResults } from "./helper.js";
+import { cacheParentResults } from "../helpers/cache.js";
 import type { CompiledQuery, InternalOptions, Adapter } from "../types.js";
 
 export type Filter = <Node, ElementNode extends Node>(
